+++ primerproyecto/src/main/java/com/miaplicacion/primerproyecto/repository/PersonaRepository.java
@@ -2,12 +2,13 @@
 package com.miaplicacion.primerproyecto.repository;
 
 import com.miaplicacion.primerproyecto.model.Persona;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface PersonaRepository extends JpaRepository<Persona , Long> {
+public interface PersonaRepository extends JpaRepository<Persona , Integer> {
